requirements.txt
Flask==1.1.1
gunicorn==20.0.4
pillow
